add a test for class planet for wave1
